reorganise and rename pages
