+++ apps/web/src/pages/about.tsx
@@ -1,4 +1,4 @@
-import { whyCreateButton, whyTextStyle } from '../../styles/why.css'
+import { whyCreateButton, whyTextStyle } from '../styles/why.css'
 import { Box, Flex, Stack, atoms } from '@zoralabs/zord'
 import Head from 'next/head'
 import Image from 'next/legacy/image'
