+++ apps/web/src/services/errorResult.ts
